requirements.txt
python-multipart==0.0.6
python-dotenv==1.0.0
fastapi==0.95.1
openai==0.27.5
pydantic==1.10.7
uvicorn==0.21.1
numpy==1.26.4
pytest==8.3.3
pytest-asyncio==0.24.0